requirements.txt
slackclient==2.7.1
python-dateutil==2.8.1